+++ src/app/components/landing/list/card/card.component.html
