+++ src/app/components/landing/list/card/card.component.ts
@@ -1,5 +1,5 @@
 import { Component, input } from '@angular/core';
-import { Card } from '../../../interfaces/card';
+import { Card } from '../../../../interfaces/card';
 
 @Component({
   selector: 'app-card',
