+++ src/app/components/landing/list/list.component.css
