+++ src/app/components/landing/list/list.component.html
@@ -3,6 +3,8 @@
     <section class="components-list">
         @for (component of components; track $index) {
             <app-card [cardInfo]="component"/>
+            <app-card [cardInfo]="component"/>
+
         }
     </section>
 </section>
